Converted code to Python v3.6 (Prev Python v2.7)
Used 2to3 tool and debugged the code till it successfully ran in Python v3.6.

Copied and pasted production code from the server which contained fixes and wasn't in the repository into the parsemail function.

Also had to adjust the check_self_for_robot_status function because it was breaking when checking emails in older formats.